feat serachStock

+++ isol/server-child/src/main/java/shinhan/server_child/domain/invest/repository/PortfolioRepository.java
@@ -7,7 +7,7 @@ import shinhan.server_child.domain.invest.entity.Portfolio;
 
 
 //@Repository
-public interface PortfolioRepository extends JpaRepository<Portfolio,Integer> {
+public interface PortfolioRepository extends JpaRepository<Portfolio,Long> {
     Optional<Portfolio> findByAccountNumAndTicker(String accountNum, String ticker);
     List<Portfolio> findByAccountNum(String accountNum);
 }
